ДЗ 1 без стилизации
